[Howling Fjord Quest - Lightning Infused Relics] The quest vehicle is having technical issues.
To be explained here. The quest Lightening Infused Relics has a vehicle specfically created for the quest. You're supposed to be to rocket jump on certain runes, and Channel the relics to complete the quest.
Well what really happens is that after completing the quests the lead to that one, the rocket jump at the camp site no longer jumps. instead you get this incredible boost that never goes up. You just look like you're going 300% speed. I feared I may get called a hacker.. but you immediately stop the crazy movement when the buff goes away.
Then at the camp site, to channel and get the relics. I was bugged out to the elevator. Any spam of the 2 button (the one that was used for the channeling) I just hopped back up and away.. every time. 

What's supposed to happen, is when you use the rocket jumping action, you go from rune pad to another to help you get around the camp site, since the vehicle itself has limited movement. Then the channeling to get the relics, you need to get close enough to it, you the quest objective would complete. 

I'm not sure if the bluff aspect was working, There was no animation, but I wasn't really being noticed by a whole lot either. So you may want to look into the whole vehicle and how it's working with it all.